+++ src/main/java/xyz/nikitacartes/easyauth/config/TechnicalConfigV1.java
@@ -22,6 +22,7 @@ public class TechnicalConfigV1 extends ConfigTemplate {
     @Deprecated
     public ArrayList<String> confirmedOnlinePlayers = new ArrayList<>();
     public transient boolean floodgateLoaded = false;
+    public transient boolean luckPermsLoaded = false;
 
     public TechnicalConfigV1() {
         super("technical.conf");
@@ -36,6 +37,9 @@ public class TechnicalConfigV1 extends ConfigTemplate {
         if (FabricLoader.getInstance().isModLoaded("floodgate")) {
             config.floodgateLoaded = true;
         }
+        if (FabricLoader.getInstance().isModLoaded("luckperms")) {
+            config.luckPermsLoaded = true;
+        }
         return config;
     }
 
@@ -46,6 +50,9 @@ public class TechnicalConfigV1 extends ConfigTemplate {
         }
         if (FabricLoader.getInstance().isModLoaded("floodgate")) {
             config.floodgateLoaded = true;
+        }
+        if (FabricLoader.getInstance().isModLoaded("luckperms")) {
+            config.luckPermsLoaded = true;
         }
         return config;
     }