+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 #### Fixes
 - Build EasyAuth against Java 17 [#199](https://github.com/NikitaCartes/EasyAuth/issues/199)
 
+#### Changes
+- Update brazilian Portuguese translation, thanks to @guigiffoni
+
 ----
 ### 3.1.11
 ##### Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2, 1.19, 1.19.1-1.19.2, 1.19.3, 1.19.4, 1.20[.1], 1.20.2, 1.20.3-1.20.4, 1.20.5-1.20.6, 1.21-1.21.1, 1.21.2-1.21.4, 1.21.5