+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ##### Minecraft 1.21.5
 #### Fixes
 - Build EasyAuth against Java 17 [#199](https://github.com/NikitaCartes/EasyAuth/issues/199)
+- Fix incompatibilities with mods by getting rid of @Redirect
 
 #### Changes
 - Update brazilian Portuguese translation, thanks to @guigiffoni